Release: RC-BETA 0.1.0 (#24)

+++ pkg/pokt/pokt_v0/models/block.go
@@ -2,5 +2,5 @@
 package models
 
 type GetLatestBlockHeightResponse struct {
-	Height uint64 `json:"height"`
+	Height uint `json:"height"`
 }
